refactor: refactor graphql code
